UIManager 수정 : Stack에서 List로 변경, 다른 씬으로 넘어갏 때 모든 UI에 접근해야됨, 다른 씬으로 넘어갈 때 UI가 Missing되는 것을 방지하기 위해 UI를 추가할 때 UIManager의 자식이 되도록함

+++ Assets/Scripts/System/Intro/IntroSystem.cs
@@ -32,7 +32,7 @@ public class IntroSystem : MonoBehaviour
             case "Google": break;
             case "Facebook": break;
             case "Guest": break;
-            default: UIManager.Instance.AddUI((int)UIPrefab.LOGIN_TYPE); break;
+            default: UIManager.Instance.AddUI(UIPrefab.LOGIN_TYPE); break;
         }
     }
 }
